moves custom hook to hooks folder
