+++ src/hooks/index.js
@@ -1,5 +1,5 @@
 // Custom Hooks import/export file
 import useLockBodyScroll from './useLockBodyScroll'
 import useRouter from './useRouter'
-
-export { useLockBodyScroll, useRouter }
+import usePosition from './usePosition'
+export { useLockBodyScroll, useRouter, usePosition }
