+++ src/pages/Home/components/MapComponent.js
@@ -1,29 +1,6 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import { Map, GoogleApiWrapper, Marker } from 'google-maps-react'
-
-export const usePosition = () => {
-  const [position, setPosition] = useState({});
-  const [error, setError] = useState(null);
-
-  const onChange = ({ coords }) => {
-    setPosition({
-      latitude: coords.latitude,
-      longitude: coords.longitude,
-    });
-  };
-  const onError = (error) => {
-    setError(error.message);
-  };
-  useEffect(() => {
-    const geo = navigator.geolocation;
-    if (!geo) {
-      setError('Geolocation is not supported');
-      return;
-    }
-    const watcher = geo.watchPosition(onChange, onError);
-    return () => geo.clearWatch(watcher);
-  }, []); return { ...position, error };
-}
+import { usePosition } from '../../../hooks'
 const positions = [
   { lat: 39.7384, lng: -105.0521 },
   { lat: 39.7487, lng: -105.0518 },
